Nest `StreamDataSource` inside `StreamDataSourceFactory`, Move File Location
